Move jvm native layer to its own header/source
