+++ imgui-binding/src/main/native/_binding.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "jni_jvm.h"
 #include "jni_common.h"
 #include "jni_callbacks.h"
 #include "jni_binding_struct.h"
