+++ imgui-binding/src/main/native/_imnodes.h
@@ -1,5 +1,4 @@
 #pragma once
 
-#include "_base.h"
+#include "_common.h"
 #include <imnodes.h>
-#include "_binding.h"
