+++ imgui-binding/src/main/native/_implot.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include "_base.h"
-#include "_binding.h"
+#include "_common.h"
 #include <imgui_internal.h>
 #include "implot.h"
