+++ imgui-binding/src/main/native/_nodeeditor.h
@@ -1,9 +1,5 @@
 #pragma once
 
-#include <stdint.h>
-#include <imgui.h>
+#include "_common.h"
 #include <imgui_node_editor.h>
 #include <imgui_node_editor_internal.h>
-#include "jni_common.h"
-#include "jni_callbacks.h"
-#include "jni_binding_struct.h"
