+++ imgui-binding/src/main/native/jni_common.h
@@ -7,8 +7,6 @@ namespace Jni
 {
     void InitCommon(JNIEnv* env);
 
-    JNIEnv* GetEnv();
-
     void ImVec2Cpy(JNIEnv* env, ImVec2* src, jobject dst);
     void ImVec2Cpy(JNIEnv* env, ImVec2 src, jobject dst);
