+++ imgui-binding/src/main/native/jni_jvm.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <jni.h>
+#include <assert.h>
+
+namespace Jni
+{
+    void InitJvm(JNIEnv* env);
+
+    JNIEnv* GetEnv();
+}
